feat: estruturar tela com UITableView e custom header/section label
